+++ x/programs/cmd/simulator/vm/genesis/rules.go
@@ -95,6 +95,6 @@ func (*Rules) FetchCustom(string) (any, bool) {
 	return nil, false
 }
 
-func (r *Rules) GetMaxActionsPerTx() int {
+func (r *Rules) GetMaxActionsPerTx() uint8 {
 	return r.g.MaxActionsPerTx
 }
